refactor(hanfts): 문서 ID 타입을 Int에서 Long(int64_t)으로 변경
문서 식별자의 범위를 확장하고 대규모 데이터셋 지원을 위해 검색 엔진 전반의 문서 ID 데이터 타입을 64비트 정수형으로 마이그레이션함.

- Kotlin의 `Document` 및 `SearchResult` 클래스의 `id` 필드 타입을 `Int`에서 `Long`으로 수정
- C++ Native 계층의 `doc_id` 타입을 `int`에서 `int64_t`로 변경
- JNI 브릿지 함수(index, remove, search, rebuild)의 시그니처를 `jint`에서 `jlong` 기반으로 업데이트
- `NativeSearchEngine`에서 검색 결과 전달 시 사용하던 `IntArray`를 `LongArray`로 교체 및 점수(score) 비트 패킹 로직 수정
- 샘플 데이터 및 `SearchViewModel` 내 테스트용 문서 ID 포맷 업데이트 (L 접미사 추가)

+++ hanfts/src/main/cpp/include/inverted_index.h
@@ -25,8 +25,8 @@ namespace fts {
 
 /** A ranked search result returned by InvertedIndex::search(). */
 struct SearchResult {
-    int   doc_id; ///< The document identifier supplied at index time.
-    float score;  ///< TF-IDF relevance score; higher is more relevant.
+    int64_t doc_id; ///< The document identifier supplied at index time.
+    float   score;  ///< TF-IDF relevance score; higher is more relevant.
 };
 
 /**
@@ -45,10 +45,10 @@ public:
      * Indexes a document, replacing any existing document with the same @p doc_id.
      * Title tokens are weighted 3× relative to body tokens.
      */
-    void addDocument(int doc_id, const std::string& title, const std::string& body);
+    void addDocument(int64_t doc_id, const std::string& title, const std::string& body);
 
     /** Removes the document with the given @p doc_id. No-op if not found. */
-    void removeDocument(int doc_id);
+    void removeDocument(int64_t doc_id);
 
     /**
      * Searches the index for documents matching @p query.
@@ -70,8 +70,8 @@ public:
 private:
     /** One entry in a term's posting list. */
     struct Posting {
-        int doc_id;
-        int term_freq; ///< Weighted frequency (title hits count 3×).
+        int64_t doc_id;
+        int     term_freq; ///< Weighted frequency (title hits count 3×).
     };
 
     /** Per-document metadata used for TF normalization. */
@@ -82,15 +82,15 @@ private:
     // Sorted map so that lower_bound() can efficiently find prefix matches.
     std::map<std::string, std::vector<Posting>> index_;
 
-    std::unordered_map<int, DocInfo>                 docs_;
-    std::unordered_map<std::string, int>             doc_freq_;   ///< Number of docs containing each term.
-    std::unordered_map<int, std::vector<std::string>> doc_terms_; ///< Reverse map: doc → terms (for removal).
+    std::unordered_map<int64_t, DocInfo>                 docs_;
+    std::unordered_map<std::string, int>                 doc_freq_;   ///< Number of docs containing each term.
+    std::unordered_map<int64_t, std::vector<std::string>> doc_terms_; ///< Reverse map: doc → terms (for removal).
 
     /** Computes the TF-IDF score contribution of a single term occurrence. */
     float termScore(int term_freq, int total_terms, int df) const;
 
     /** Accumulates weighted token frequencies into @p term_counts. */
-    void indexTokens(int doc_id,
+    void indexTokens(int64_t doc_id,
                      const std::vector<std::string>& tokens,
                      int weight,
                      std::unordered_map<std::string, int>& term_counts);
